Fix : Issue with multipart and its response

+++ src/main/java/com/pop/codelab/chatopbackend/user/UserRepository.java
@@ -4,8 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
   Optional<User> findByEmail(String email);
+
+ // Optional<User> findUserById(Long id);
 
 }
